pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dw6"
version = "0.1.0"
dependencies = [
    "python-dotenv",
]

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-cov",
]
